refactor: use signal instead of member function to handle signals
